pyproject.toml
[tool.poetry]
name = "TestCapitalCom"
version = "0.1.0"
description = ""
authors = ["Alexander Tomelo <qa.tomelo.an@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "7.4.2"
webdriver-manager = "4.0.1"
selenium = "4.15"
pytest-html = "4.0.2"
allure-pytest = "2.13.2"
pytest-xdist = "3.3.1"
#pyautogui = "^0.9.53"
psutil = "^5.9.4"
#aiogram = "^2.24"
#black = "23"
flake8 = "6.1"
#ruff
#memory-profiler
speedtest-cli = "^2.1.3"
qreader = "^3.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)s] %(message)s (%(filename)s:%(lineno)s)"
log_file = "pytest.log"
log_file_level = "DEBUG"
log_file_format = "%(asctime)s [%(levelname)s] %(message)s (%(filename)s:%(lineno)s)"
log_file_date_format = "%Y-%m-%d %H:%M:%S"
